Split cosign/lib.rs into smaller files
The `cosign/lib.rs` file was too big. This is now splitted into files
that have a clear purpose.

This makes navigating and maintaining the code easier.

While doing that, some of the inline documentation has been moved from
the specific struct into the top level library.

For example, examples about how to use `cosign::Client` have been moved
into the top-level documentation.
As a result of that, examples about how to use the library are
immediately shown on the main page of the library. This improves the
whole developer experience.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>