fix: filter_signature_layers() when no match
We iterate through all constraints, and filter by the constraint. A
passing verification (bool true), means that it ends in the resulting
vector.

Remove negation inside of the `match` for constraints. This means that
verified layers will get correctly passed to `any()`.

Remove negation on the `constraints.iter.any()`. this means that when a
layer is verified (returning true in the closure), it well get added
correctly to the layers vector.

All in all, this gives us the correct behaviour: if there's some
verified signatures, they get added.
If there's no verified signatures, we return
`SigstoreError::SigstoreNoVerifiedLayer`. In the past, we where almost
always returning that error instead.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>