feat: Add `SigstoreVerifyConstraintsError{ unsatisfied_constraints }`
which contains a vec of refs to unsatisfied constraints. This will be
returned by verify_constraints().

It is not part of `SigstoreError` because it contains a Vec of Box dyn,
which means it needs lifetime annotations, and it would force us to add
lifetime annotations to all `SigstoreError` and their consumers.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>
Co-Authored-by: Flavio Castelli <fcastelli@suse.com>
Signed-off-by: Flavio Castelli <fcastelli@suse.com>