cosign client: introduce caching option
Allow end consumers of the library to create a `cosign::Client` that
caches the response it gets from OCI registries.

The cache entries are automatically expired after 60 seconds.

This can be useful in certain scenarios, like when writing code
that can verify the same container image, but with different
constraints.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>