Improve public key output and add file output
Previously Debug printing was used to output the public key after
retrieval. Using Display printing the public key is displayed to the
command line as expected with line breaks.

In addition this commit adds --output,-o as arguments to store the
retrieved public key in a file.
To ease use a default value (key.pub) is provided, so that no parameter
needs to be given.

Resolves: #105
Signed-off-by: Gronner <felix.braeunling@t-online.de>