Changed the type of LogEntry.body from String to Body
Objective
To enable detailed information of LogEntry.body to be obtained when LogEntry is retrieved.
Change
In conjunction with changing the type of LogEntry.body to Body, I added default trait for Body and the structs and enums on which it depends, using derive macro.
 (Body, Spec, Signature, PublicKey, Data, AlgotithmKind, Hash)
In addition, in entries_api.rs, the use of serde_json::from_str was changed to the newly implemented LogEntry::from_str.
LogEntry::from_str uses HashMap with &str for key and serde_json::Value for value, and body is decoded.
Since a HashMap cannot be directly converted to a LogEntry, it is converted to a String with serde_json::to_string and then converted to a LogEntry with serde_json::from_str.

Signed-off-by: Neccolini <shun11202991@gmail.com>