automation: fix GHA invoking cargo audit
We recently got a major version update of the audit-check GHA.

This version invokes:

```console
cargo audit --json --file ./Cargo.lock
````

Which fails, because we don't have a `Cargo.lock` file.

This commit makes sure the `Cargo.lock` is present prior to
invoking `cargo audit`.

While fixing this issue, I've also consolidated the two different `cargo
audit` GHA we had into a single one. It was just a matter of handling
cron and push events inside of the same GHA.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>