chore: Add feature test Makefile targets
This adds check-features, check-features-native-tls and
check-features-rustls-tls targets to the Makefile to allow local testing
of features.

Given that these tests take a long time to run, they've been added as 3
seperate targets. Most issues are likely to be caught in CI, but this
allows a contributor to easily reproduce and fix these issues in their
own development environment.

The targets also require that the user has installed cargo-hack, but the
error message from cargo when this is missing is pretty informative.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>