ci: Add TOML format checks
This checks in a .taplo.yml file that was used to format the Cargo.toml
files with the recent changes and adds a check in CI to ensure that all
files are formatted the same way.

In addition, it changes all remaining uses of installing rust using
rustup directly to instead use the dtolnay/rust-toolchain action.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>