chore(deps): do not require ring
Replace the ring dependency with `aws-lc-rs` whenever possible. The
`ring` dependency is used only when building for the
`wasm32-unknown-unknown` target, since `aws-lc-rs` doesn't support it.

Note: `ring` is still shown inside of the development dependency tree,
since the `testcontainers` crate is still using it. However, the `main`
branch of `testcontainers` has already made that optional.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>