fix!: handle multiple Rekor keys
Handle multiple Rekor keys found inside of Sigstore's TUF repository.

At bundle verification time, the proper key is found by looking up the
key using its ID.

Inside of `trusted_root.json`, each key has a `logId.keyID`, which is a
base64 encoded value.
This value must be base64 decoded and then encoded using hex format to
build the key ID.

The bundle has `logID` attribute, that is the ID of the key that was
used to sign the bundle. This ID must match one of the hex-encoded
values built before.

This commit breaks some public APIs of the library, causing us to make a
minor release.

The examples have been updated to reflect the new API changes.

The main example about verification (the one bundled inside of our docs)
has been changed to build the trust root from the contents of the TUF
repository, instead of using local data. This is a more common usage
pattern of the library.

Fixes #501

Signed-off-by: Flavio Castelli <fcastelli@suse.com>